feat: init atwho autocompleter in pull request comments
Part of story #38643 @ notifications in pull requests

How to test:
- Create a pull request
- When you edit the description of the pull request, or when you add or
edit a global comment, or when you add or edit an inline comment in a
file, when you type @ and 3 characters, the autocompleter should appear
to suggest matching users.
- No other functional change.

Why?
Add @tuleap/mention as dev dependency of pullrequest-overview, because
otherwise I cannot mock it in tests. It depends on jquery being a global
variable, and we want to avoid pulling it in tests "just" for that.

Add code to dispatch an "input" event after inserting the username, so
that the hybrids code that depends on input events can catch the change
in the textarea, even when it's the last thing the user writes.

Change-Id: I26c8e7615942e3f32bbbd5b38fe6cf16906d3993