fix: request #41115 Cannot filter artifacts with an open value of an open list via the REST API
Until now it was possible to filter using values coming from a bind
value but not one of the open values (i.e. one of the values not preset
when the field is created). The open values used the same IDs range than
the static bind values so it was not possible to distinguish them.

The ID of the open values can now be distinguished and used to filter
the artifacts.
This is not a breaking change of the API:
* IDs of the open values have changed but they were not usable anyway
* IDs of the open values are now a string but it is already the case of
  the bind values when user groups are used

Change-Id: I2eb1dc1f24b46d24d0d12e7e91ffdf83d7002671