fix: alignment of freetext sections in readonly
When a freetext section is displayed on a document, when the user does
not have permissions to update the document, the content of artifact and
freetext sections are not aligned.

This is due to the fact that since there is no swatch color applied to
the freetext section the border is not displayed at all. By forcing the
color to be `transparent` the border is the same for every sections.

Note: this is the same for artidoc in write mode, but this will be
handled later.

Part of story #41105: Have freetext section in artidoc

Change-Id: I1e0e29c079f7845623a4f89d9da01ed80d056c8a