feat: Focus should work with keyboard
Part of story #38643 @ notifications in pull requests

How to test:
In pull request comments (either in overview or in file diff page), when
you use the tabulation ⭾ key to focus the textarea, the left border and
the tabs (Write / preview) should be correctly colored, the same way as
when you click in the textarea.

Why?
No need to listen to events in the big document, we can "simply" listen
to "focusin" and "focusout" events on the element itself. It will tell
us whether something inside is focused.

We also don't need to programmatically blur the textarea, because it is
already NOT focused in the "focusout" event and when we call
`controller.resetWritingZone()`, either the element is disconnected from
the DOM (so the textarea can never be focused) or the user just clicked
on the "save" button (and that button is focused).

Change-Id: I8b1f6f5a7139cbb6b2636487abbc45eb603a551b