chore(forge-config): Remove PluginFileInfo where it's useless
It seems that developers have copy/paste plugins base and inherited
PluginFileInfo without having configuration. It's been removed.

DynamicCredentials had a mechanism introduced in change
I387f670ebb68c50acec48608220d45bc65e4a6ab to fallback on file base
configuration after the move to database. As we are the only one to use
this plugin and everything is now deployed with `config-set` we can
remove the fallback mechanism.

TrackerDateReminder had a log mecanism but as the config file was named
"config.ini.dist" it was created as "config.ini" and thus not loaded
(only .inc files are taken into account). I choose to remove the code
entierly because it's tracker v3. Note: I was not able to test my change
as we can no longer create tracker v3.

Part of: request #40460 Have a single source of truth for configuration variables

Change-Id: I668f8c40b5351649247f515a38d9590039f582b8