fix: import existing with misconfigured tracker
When the artidoc is configured with a tracker without title semantique,
the import of existing section is weird: it does nothing.

Now an error message is displayed instead of a dummy form.

Took the opportunity to restrict the check only on the title field (we
are searching artifact by title, no need to restrict the search of
existing artifacts if the description is not submittable).

Took the opportunity to adjust the title of the modal to match the label
of the dropdown menu that triggered it (Import existing requirement
instead of Import existing section).

Part of story #41105: Have freetext section in artidoc

Change-Id: I16870a8a1fba5000a777dc7afbb6266884ef7536