git repositories-list replace indexed array of repositories by an object by object
part of request #31168 Migration Vue2 => Vue3

Why?
When we bump to vue3 (gerrit #32664) we have an issue `repositories_for_owner` was built like this:
[
   0: {id: 102, uri: "/users/102", user_url: "..", realname: "...", display_name: "..."}, // this is a user who have forked the repository
   project: [
      {id: 1, label: "my repo" ...} // this is a git/gitlab repository
   ]
   102: [
      {id: 1, label: "my repo" ...}  // this are the forks of my user
   ]
]
With the bump the key (0|project|102) are ignored, the selectbox for forked repository no longer works

I do not understand why we originally mix users and repositories and it's crap, this patch fixes most of the update issues when we migrate to vue3
But when we are in alphabetical order, the fork user selectbox still does not work.
I'll extract the users from repository list in a dedicated patchset

The indexed array have been transformed into an object so we can use regular typescript code to crud our data:
```
export type RepositoriesForOwner = {
    id: number | string;
    repositories: Array<Repository | FormattedGitLabRepository>;
};
```

No functional changes:
 - git and git lab repositories are displayed
 - when a gilab repository is unlinked, the repository list is up to date
 - we can switch display mode to alphabetical/path
 - repositories can be filtered
 - selectbox for fork is ok

Change-Id: I1727b134407c21f6e56219738b772d4e6326554a