feat: introduce section version table
This commit prepares the field to be able to store different version of a
document so that we will be able to retrieve a former version of a
freetext section (artifact section content is defacto versionned thanks
to artifact changesets).

No functional changes.

Part of story #41105: Have freetext section in artidoc

Change-Id: I64908f8d498f367b80637329953a7330eaf4b7d6