chore: Split user notifications preferences
Part of story #38681 New tracker notification control

As we will bring new controls to notifications preferences, we first
need to split to let Trackers have its own section.

Note: the code displaying this page is in core, so I choose to use an
event catch by plugin Tracker to then inject in mustache template. This
is not perfect but the html returned by the plugin comes from mustache
so it avoid malicious injections.

*Testing:*

You should be able to change your trackers notifications preferences and
they should be applied.

Change-Id: I7a07da12ddb23a43c2b14716074acbc82e235223