refactor: Composition API in Iterations section
Part of request #41095 Migrate plugin-program-management-iterations to
Vue 3

When browsing a PI, no functional change in the iterations section (the
right part). It should still load the iterations of the current PI, show
a skeleton, and show an error message with the iteration custom labels
(you can throw a RestException in
\Tuleap\ProgramManagement\REST\v1\ProgramIncrementResource::getIterations)

Why?
I removed the v-if from <iteration-card> because the linter complained
about using both v-if and v-for. It should work the same way (the list
is empty while loading and when there is an error).

I removed the computeds because the iteration labels, program_increment
and iteration_tracker_id cannot change during runtime, they are injected
once in the mount point and are readonly afterwards. We only need to
compute them once during component setup.

Change-Id: I7763ca5b39fb3a6ed6964d649c6ce47028bcafa4