refactor: Composition API in IterationUserStoryList
Part of request #41095 Migrate plugin-program-management-iterations to
Vue 3

When browsing a PI, when you unwrap an iteration on the right part, no
functional change in the list of user stories. It should still load the
user stories linked to the iteration, and show an error message if it
fails (you can throw a RestException in
\Tuleap\ProgramManagement\REST\v1\IterationResource::getIterations,
which should be named getContent)

Why?
I removed the v-else from <user-story-card> because the linter
complained about using both v-if and v-for. It should work the same way
(the list is empty while loading and when there is an error).

Change-Id: I763abd4084666d2dd4f286a35c948e75e4694815