feat: Add a new checkbox in user notifications preferences
Part of story #38681 New tracker notification control

This new checkbox will allow user to disable mailing for all updates on
artifacts. The preference value is not yet used to filter mail
recipients

*Testing:*

You should be able to change state of this checkbox. In database it
should be stored in table `user_preferences` with the name
`user_notifications_all_updates_tracker`

Change-Id: I415e3fb4bdc9b6308d83ffafa622841234bcd2d1