refactor: Composition API in AppBreadcrumb
Part of request #41095 Migrate plugin-program-management-iterations to
Vue 3

How to test:
When browsing a PI, no functional change in the Breadcrumb.

Why?
Renamed Breadcrumb to AppBreadcrumb (vue/multi-word-component-names).

For program_flags, we cannot go through `useNamespacedState`. It
encapsulates the state in a `readonly()`, but Vue 2 does not support
readonly for arrays. The only solution is to access the state from
`useStore` directly, to avoid the `readonly()` array. This solution is
not intended to stay as it is, it should be changed after the switch to
Vue 3.

Change-Id: If94309b0de1d2ae5bcc27a1dcedd4927b82a77ab