Convert SwimlaneSkeleton into composition API
part of request #41443 convert taskboard into vue3

Why?
Class components are no longer supported in Vue 3. We must switch to composition API to upgrade to Vue 3.

For columns state, we cannot go through `useNamespacedState`. It
encapsulates the state in a `readonly()`, but Vue 2 does not support
readonly for arrays. The only solution is to access the state from
`useStore` directly, to avoid the `readonly()` array. This solution is
not intended to stay as it is, it should be changed after the switch to
Vue 3.

No functional changes

Change-Id: I1fb89624a68693956834b61ecff393a797c7ed2d