refactor: do not depend on PFUser DB methods for "On all updates"
Part of story #38681 New tracker notification control

How to test:
No functional change expected when enabling or disabling the new
checkbox in user preferences > Notifications > Trackers section.

Why?
PFUser has many methods that access the database "on their own" by
instatiating their own DAO. Seeing as someone already did the work of
writing an interface for that DAO, we should go through it. As a result,
our methods are no longer static (but since they handle DB storage, they
probably should not have been). The "Preference" is extracted as a plain
old object, to clarify that it's not just any boolean. We can build it
from the HTTPRequest, and from the DB in Retriever.
The "on own action" preference will come in a next commit.

Change-Id: I454bf4dab04ec54f646b5ded82108181ae5ee2cb