refactor: Composition API in IterationsToBePlannedSection
Part of request #41095 Migrate plugin-program-management-iterations to
Vue 3

How to test:
When browsing a PI, no functional change in the list of unplanned user
stories (on the left). It should still load the user stories that are
not planned yet, and show an error message if it fails (you can throw a
RestException in
\Tuleap\ProgramManagement\REST\v1\ProgramIncrementResource::getBacklog).

Why?
I removed the v-else from <user-story-card> because the linter
complained about using both v-if and v-for. It should work the same way
(the list is empty while loading and when there is an error).

Change-Id: I3aeefb380c77124040ed4392ccf8e560e1eb3bd7