refactor: do not depend on PFUser DB methods for "On own action"
Part of story #38681 New tracker notification control

How to test:
No functional change expected when enabling or disabling the "on own
action" checkbox in user preferences > Notifications > Trackers section.
No functional change expected on handling that preference: if the
preference is disabled, user should not receive e-mails when writing
comments.

Why?
PFUser has many methods that access the database "on their own" by
instatiating their own DAO. Seeing as someone already did the work of
writing an interface for that DAO, we should go through it. As a result,
our methods are no longer static (but since they handle DB storage, they
probably should not have been). The "Preference" is extracted as a plain
old object, to clarify that it's not just any boolean. We can build it
from the HTTPRequest, and from the DB in Retriever.

Change-Id: I0dfc10607fc7ef3ada50a0359c17daa186984d2b