feat: Introduce TUS implementation based of UUID
Current TUS implementation is based on autoincremented integer id. We
should use UUID instead.

A dedicated next-generation implementation is provided (exactly the same
as before but with UUID) so that we have time to convert existing code
that still rely on integer ids.

No functional changes.

This patch is provided as a standalone to not produce a too big review.
Therefore there is nothing much to test.

Part of story #41105: Have freetext section in artidoc

Change-Id: I2731d1e452dadf619f5b34155dbccad2199e3ced