refactor: replace translate directives by gettext
part of request #41457 Migrate plugin-program-management-backlog to Vue
3

No functional change expected.

Why?
The v-translate and translate Vue directives are deprecated and removed
in Vue 3. This is mandatory for the conversion.
I chose to remove all snapshot tests that broke.They are more
effectively replaced by limited assertions that will not break when we
change a CSS class or a translated string.

Change-Id: Icf8ee01221ad15d171fe8e7256dcd3530384b4a4