feat(forge-config): import docman config to db
Move variables to ForgeConfig.

This commit is a bit large as it introduce a common infrastructure to
easy the writing of migration buckets. Unlike git where things where
quite simple (strings or int) in docman we have booleans that can be
represented by all the values in php (true,false,1,0,...) but that
should be stored as string in the forgeconfig database.

As we needed a more formal description of the variables to be imported,
I choose to leverge classes and types to help writing code that deal
with all those conversions.

Closes request #41447 Import Docman configuration variables in DB

Change-Id: Ifac52754a8437ff74baa87f2c8f1fa543addb437