feat: Delete dangling files
Artidoc upload files have an expiry date. If the upload is not
finished at this date, then they are removed.

To test, call the API REST to create two *empty*(1) files.
POST artidoc_files
{
  "artidoc_id": 636,
  "name": "filename1.png",
  "file_size": 0,
  "file_type": "image/png"
}

POST artidoc_files
{
  "artidoc_id": 636,
  "name": "filename2.png",
  "file_size": 0,
  "file_type": "image/png"
}

In the table plugin_artidoc_section_upload you should see the two
entries. On the filesystem, you should see two corresponding files in
/var/tmp/artidoc/ongoing-sections-file-upload/.

Update the entry for filename2.png to set an expiration date in the past
(example: 1234567890).

Now execute `tuleap daily-job` on the server.

=> In the table plugin_artidoc_section_upload filename2.png is gone and
the corresponding file in /var/tmp/artidoc/ongoing-sections-file-upload/
is also removed. No change for filename1.png.

Note:
(1) Current contribution is not bound to empty files, but testing with
empty files is the fastest way to contribute relatively small patches
for this. At the end of the story, empty files wont stay in the
temporary upload area.

Change-Id: Ia67b6b63cdf5499708033f7f5048a7df93a83c41