Remove un needed mixin for collapse column handling
part of request #41443 convert taskboard into vue3

Why?
Class components are no longer supported in Vue 3. We must switch to composition API to upgrade to Vue 3.

Note:
According to vue documentation we should switch from mixin to composable in composition API [0],
In this particular case, the mixin was written only arround the `expandColumn` actions
Even if it works, intention of code is clearer when it's directly done in the actions
It also enable us to have more consistent unit test :
 - the if is tested directly in action
 - in component we only need to check that the proper action is called and not its behaviour)
I needed to use the `override` arround column props, because it's also declared in other mixins ...

I'll convert them later in order to keep this change small

How to test:
 - you can expand/collapse columns, it is stored in you user preferences (same as reload)
 - you can drag and drop cars in expanded/collapsed columns
 - you can expand/collapse invalid mapping cells

[0] https://vuejs.org/api/options-composition#mixins

Change-Id: I88ae606e3e61db90241a0c85dde3905c4696611e