refactor: extract loading from SectionsStore
part of request #41467 Refactor artidoc's composable architecture

The goal is to split the sections store to make it become a simple
SectionsCollection, and take down piece by piece the various composables
messing with the app's data flow.

This first step allows us to simplify a lot of checks here and there to
make sure the sections exist for the current document.

-- no functional change expected

Change-Id: Ied7d770b6f62b8d34b88a617de56753e08f9fdf9