Add the XSLX export button
Part of story #41463 XLSX export

Note: I prefered to duplicate the `ExportCSVButton` to avoid having to deal with the case where we are in a expert mode or default mode context.
It will also minimise the chance of regression and make easier to deal with `ExportCSVButton` component when the default mode will be removed.
Potentially... The name of componenent does not match anymore

How to test:
 - In your XTS widget go to the expert mode (if it is not already done)
 - Open your browser console
 - Click on the new `Export XSLX` button
=> the console log is displayed

Change-Id: Ia5f0f27ec2ca80a7864c617171bf5d6f3aae54cf