fix: push unsaved comment to WritingZone
fix: request #40100 Pull Request comment is empty after scrolling

How to test:
- No functional change when editing a pull request's description (first
comment).
- Craft a pull request with a file > 200 lines. When you browse that
file in the PR diff and you write a new inline comment, if you scroll up
or down about 100 lines, and then go back to your comment, the unsaved
text you wrote should still be shown. When you save it, the comment is
created with your text.
- When you edit an existing inline comment, and you scroll up or down
about 100 lines, and then go back to your comment, the unsaved edited
text you wrote should still be shown. When you save it, the comment is
edited with your text.
- No functional change for global comments (in the PR overview tab).

Why?
Passing unsaved content to the WritingZoneConfig (See [0]) made Hybrids
re-build it for each typed character, which is both harmful to UX (you
cannot edit anything in the middle of text as the caret is yanked to the
end) and inefficient. It also caused a flickering effect on the left
border of the comment, due to it being re-created constantly.
It was still better than nothing because at least unsaved data was not
lost.

In this change, we use a different approach. The comment string under
edition is now a property of the WritingZone element. Callers can set it
with the unsaved content. In order to avoid another loop, the
writing_zone and writing_zone_controller properties **must not** depend
on the unsaved content. They must not "pull" the presenters, otherwise
Hybrids will re-create them for each typed letter.
Therefore, we switch to a "push" update model: whenever the presenters
(containing the comment string) are saved in the three types of
comments, their setter function also sets the property on the writing
zone. Thus, when the WritingZone is re-created because of scrolling in
CodeMirror, it starts with the correct comment in its textarea. Then, we
make sure to keep this property properly up-to-date with what the user
is writing in the textarea.

I chose to remove the ElementContainingAWritingZone type because using
it meant WritingZone can affect its parent controller. From a code
design point of view, it's simpler when it's the parent's responsibility
to add the event listener and set the controller on the writing zone,
because it puts fewer constraints on the parent. It should not be
required to have a controller, or to listen to the event. It's one of
the few cases where I think it's better to have "duplicated" code.

I also chose to move presenter assignations to the property descriptors
when they were in controllers. When the controller "just" assigns a
property to its host, in my opinion it should be the role of the element
and its property descriptors. The controller should be responsible for
calling the API and applying the business rules of the element. It
should not depend on the element host, so that UI and business rules can
be kept separate.

[0] 7683d45314b68031d5cef90be5e276824dc520ba

Change-Id: I6a4bb46ac487bbbf41c7fd49703183e37965481c