refactor: extract file-upload library
prose-mirror-editor is starting to polluting everywhere in artidoc. In
order to reduce the coupling, the file upload part of
prose-mirror-editor (to be able to drop image) should be extracted in a
dedicated library.

```
artidoc --------------------、
  |                          \
  `--> prose-mirror-editor ---`--> file-upload
```

This contribution extracts only the types for now. At the end
file-upload will handle the full upload via TUS protocol.

Part of story #41105: Have freetext section in artidoc

Change-Id: Ie026bb485a74bd98ca2d057f1083f7c0545b9e23