fix: request #39753 Migration should not fail when MW Standalone was installed directly in 1.39
When migrating from the MediaWiki plugin to the MediaWiki Standalone
plugin the migration process does something like this: MW 1.23 -> MW
1.35 -> MW 1.39 because jumping from 1.23 to 1.39 directly is not
possible.

In order to do that MW 1.35 should have a valid configuration file so
the general configuration of MW Standalone is used. This is problematic
for instances that a MW 1.23 to migrate but only installed the MW
Standalone after we bumped it to 1.39 because a new theme was added
(which is unknown to MW 1.35). MW 1.35 will try to load a theme it does
not know and it will fail preventing the migration to be succeed. In
order to fix that, we now update the automatically generated
configuration file to not load unnecessary skins.

Change-Id: Ibfacee2354720a379758b9161888d01168a5be54