refactor: extract editors states into a global collection
part of request #41467 Refactor artidoc's composable architecture

In this contribution, the editors states managed by the useSectionEditor
composable are extracted from it and made available to all the
components through an injected collection of states.

Note: Sorry for the big patch. Its size shows how much the composables
mess with the app's architecture.

-- no functional change expected

Change-Id: I9fe3334d235dfe544b1b35361143d8f4a632493a