refactor: Move CurrentArtifactIdentifier to common lib
part of story #41466 Have new Artifact Links in Artifact View

No functional change.

Why?
The goal of this series of refactorings is to extract the Artifact Links
field from the Artifact modal to its own library, so that we can reuse
it in the full-page Artifact view.
In order to do that, we first need to move common types and classes to a
third library. In order to do its job, the link field needs to know a
couple of things: what is the current artifact ID, what is the parent
tracker ID, etc. It also needs to dispatch events, for example to show
REST error messages, or to disable the submit button. Since this code
belongs to the artifact modal domain, and we do not want to have a
dependency from the Artifact view to the modal, we must move it to a
third library. Later, both the link field and the modal itself will
depend on the "artifact-common" library.

Change-Id: Ib3ba6a24e7cdb16140ae722a4f40174d9a9be081