feat: TLP dropdowns can contains actions (filter, CTA)
Until now it was only available in git, but we can put filters or CTA in
a dropdown.

Specific code in git is moved to TLP standard. (No functional changes)

An example is given in TLP storybook to demonstrate how filter and CTA
can be added inside a dropdown menu.

This will be necessary to implement the query selector in cross tracker
search widget.

Part of story #41799: Have multiple queries per widget

Change-Id: I3f59e0b030b626be56e4c368fbc55be41612a9c5