refactor: Move EventDispatcher to common lib
part of story #41466 Have new Artifact Links in Artifact View

No functional change expected.

Why?
The goal of this series of refactorings is to extract the Artifact Links
field from the Artifact modal to its own library, so that we can reuse
it in the full-page Artifact view.
In order to do that, we first need to move common types and classes to a
third library. In order to do its job, the link field needs to dispatch
events, for example to show REST error messages, or to disable the
submit button. Since this code belongs to the artifact modal domain, and
we do not want to have a dependency from the Artifact view to the modal,
we must move it to a third library. Later, both the link field and the
modal itself will depend on the "artifact-common" library.

Events need a trade-off. We can keep the full list of events fixed and
known at typecheck time, but it means the other fields must leak into
the common library (we must know about File upload and list fields,
because there are events for them). I chose a second solution: making
the full list of events unknown and generic. It means autocompletion is
weaker, we are not warned that we are observing or dispatching an event
that does not exist. We still get an error about accessing a property on
an undeclared event, but the error will not tell us that we need to
declare the event…

Change-Id: I925571007d04969c6702447782a3dbc8f8a6a753