Get ride of default mode in frontend
Part of story #41464 Convert non expert queries

The last blocking point was the ability to export the query result in a file.
Now it possible to export query result, we can burn the default mode and say so
long...

`getTextResponse` is was also removed from fetch-result library since it
was to retrieve the CSV exported in the default mode

Expert mode should work as before.

Change-Id: I2802483f127f9f99df8228fe515db0a9b5bdb7a5