refactor: Composition API in ErrorDisplayer
part of request #41457 Migrate plugin-program-management-backlog to Vue
3

How to test:
No functional change expected. When browsing a Program, hack the PHP
resource class so that it sends a REST error. For each case, you should
see an error alert with a translated text:

\Tuleap\ProgramManagement\REST\v1\ProjectResource::getBacklog() => in
the "To plan" part on the left

\Tuleap\ProgramManagement\REST\v1\ProgramBacklogItemsResource::getChildren()
=> when you expand the children of a Feature card

\Tuleap\ProgramManagement\REST\v1\ProgramIncrementResource::getContent()
=> when you expand the Features planned in a Program Increment

Change-Id: I342558f11f67df8c29c287228124091182085328