refactor: Composition API in ErrorModal
part of request #41457 Migrate plugin-program-management-backlog to Vue
3

How to test:
No functional change expected. Hack the PHP resource class
\Tuleap\ProgramManagement\REST\v1\ProgramIncrementResource::patchContent()
so that it sends a REST error.

Browse a Program Backlog, expand a Program Increment card on the right.
Select a Feature card in the Backlog on the left and try to drop it in
the open Program Increment. You should see an error modal open. You can
open the details to see the error message you entered in the REST error.

Change-Id: Ia047c8bb501757bbefd7f306f87706cb3602cb95