chore: request #41797 cypress: 13.15.0 -> 14.0.1
No functional change, make tests-e2e should pass.

If you plan to run tests with make tests-e2e-dev, you should first run
`cypress install` which is not done automatically anymore.

Why?
XTS: wait for table header to be updated by Vue, otherwise Cypress reads
the cells from the previous query.

TQL and XTS: For some reason, I could not type() into the CodeMirror
editor. I found another selector based on attributes (role and
contenteditable), so I replaced all usages of .cm-editor. text() is
faster anyway and correctly sets the CodeMirror with our queries.

Document: wait for skeleton to disappear (because we need to wait for
loading to be complete).

Kanban: wait for spinners to disappear (because we need to wait for the
columns to be loaded, otherwise the data is wrong. I also removed some
computed data-test attributes, we can use the visible column label
instead, the commands are barely longer.

Change-Id: Ibb921d2b8becaf9525bf9e877d7f26476ccb25a0