refactor: Composition API in ProgramIncrementFeatureList
How to test:
When browsing a Program, when you have the correct permissions, you can
drag/drop Features between the Program Increments and the Backlog. If
you don't have permission to update the Program Increments Artifact link
field, the cards in the PIs cannot be moved, and you cannot drop a card
from the Backlog in a PI. The list of Features in a PI is cached and
fetched only once.

Why?
I removed the v-else from <feature-card> because the linter
complained about using both v-if and v-for. It should work the same way
(the list is empty while loading and when there is an error).

Change-Id: I10184c7a119b37757339c49a7147ecae035113e0