refactor: Composition API in ProgramIncrementFeatureList
part of request #41457 Migrate plugin-program-management-backlog to Vue
3

How to test:
When browsing a Program, you see the list of Program Increments on the
right. They can be opened/closed. If the configuration defined a tracker
for Iterations, there is a link to the "plan iterations" app with the
chosen label.

Change-Id: I873a671cd8fb2bcee857c7663180988a61ce63b7