refactor: Composition API in ProgramIncrementList
part of request #41457 Migrate plugin-program-management-backlog to Vue
3

How to test:
When browsing a Program, you see the list of Program Increments on the
right. If you have the right permission, you can create a new PI with
the button (it redirects to create artifact view). The custom labels
defined in administration are used.

Change-Id: Ieaa351f3bdaa4374579821529030aff0d2a77bcd