refactor: Copy CreateArtifact interface to the link field
part of story #41466 Have new Artifact Links in Artifact View

No functional change.

Why?
The goal of this series of refactorings is to extract the Artifact Links
field from the Artifact modal to its own library, so that we can reuse
it in the full-page Artifact view.
CreateArtifact was being shared by the big Modal (when you create a new
artifact) and by the Links field (when you create an artifact with just
a title, to link it directly). Instead of moving this to the "shared"
library, I chose to have two implementations. In Links field, all we
need is the title, and as far as we know, we are not going to add
anything else to this artifact. In fact, we even require from the
tracker that it allows creation with just a title. So in my opinion, it
makes sense to have a specialized interface "just" for that. It avoids
adding more grey zones and blurry types: in the context of the big
modal, we don't know exactly all the fields at typecheck time, in the
Links field we know there is exactly one title.

I also took the opportunity to simplify the identifier creation by
"just" having a constructor from a number, instead of a Proxy +
interface. This is an application of ADR-0005 in Program Management [0],
which is also valid in the artifact modal as some concepts have been
used here too.

[0]
https://tuleap.net/plugins/git/tuleap/tuleap/stable?a=blob&hb=fcd7aaac3fd5392ead09280ad01a33096b2a5681&f=plugins%2Fprogram_management%2Fdocs%2Fdecisions%2F0005-build-domain-objects.md

Change-Id: I100bee8661cfa78fdb35f1d67aee2e24c2470c58