feat: introduce level for sections
Sections now have a level. For now it is not visible on the user
interface, only in REST representations of the section. For example when
you load an artidoc, you can see in the dev console that the REST call
returns `level: 1` for every sections. If you hack the database to set
level=2 for a given section, then it will be reflected in the next
loading of the artidoc.

Part of story #41805: Management of titles

Change-Id: Ib5da4fd1621a362a38aa86839bae3913bf74bd4a