fix: Display errors in link artifact creation
fixes request #41806 Errors are not shown when creating an artifact in the
Links field of the modal

How to test:
Open the artifact modal (for example Kanban), create an artifact and
link it directly from the Links field. When there is a problem during
trackers retrieval, projects retrieval or artifact creation, the error
message should be shown in the Links field (and the modal scrolls to
it).

The error was not displayed,
probably caused by f755f7e1d20aaf2ad392c900b26874cd383c4e36.

Change-Id: I54f097136ddc6fb955ab2bc03ce21898e0560ba0