feat: request #41809 Send PHP FPM logs to the system logs instead of a specific file
Note this will only be really effective on fresh instances or with the
next PHP upgrade as Tuleap does not rewrite the included log files.

Note the default global PHP FPM config is now rewritten to be sure all
logs ands up where we expect them to go and to make sure PHP FPM does
not complains if the log folder is not present on the disk.

Change-Id: Ib36cadf2ca28e8266efc4cc2ff9af3663b78459e