feat: display section according to the level
Part of story #41805: Management of titles

This commit manages the display of numbers in front of sections titles, in the table of contents, in print version and in front of sections in artidoc (in readonly mode and in edit mode).
When reordering sections, numbers are updated.
The management of numbers when adding new sections will be done in dedicated commits.

Please note:
- Section numbering used to be displayed using `<ol>` and `<li>` tags, but with the addition of sub-sections, they are now calculated according to level.
- If the template used manages numbering itself, you risk ending up with things like `1.2. 1.2.`. To avoid it, I replace the `<ol>` `<il>` by `<div>`.
- To be consistent, the display level ends with `. ` everywhere (in TOC, PDF and also in front of sections).
- The placement of the numbers in front of the sections has been calculated to be aligned with the bottom of the titles.

How to test:
Change section levels in the database

Change-Id: Id2fa71c5923ca33447cf5fa271040ce44495141e