refactor: Move ParentArtifactIdentifier to common lib
part of story #41466 Have new Artifact Links in Artifact View

No functional change.

Why?
The goal of this series of refactorings is to extract the Artifact Links
field from the Artifact modal to its own library, so that we can reuse
it in the full-page Artifact view.
In order to do that, we first need to move common types and classes to a
third library. In order to do its job, the link field needs to know a
couple of things: what is the current artifact ID, what is the parent
tracker ID, etc. Since this code belongs to the artifact modal domain,
and we do not want to have a dependency from the Artifact view to the
modal, we must move it to a third library. Later, both the link field
and the modal itself will depend on the "artifact-common" library.

I also took the opportunity to simplify the identifier creation by
"just" having a constructor from a number, instead of a Proxy +
interface. This is an application of ADR-0005 in Program Management [0],
which is also valid in the artifact modal as some concepts have been
used here too.

[0]
https://tuleap.net/plugins/git/tuleap/tuleap/stable?a=blob&hb=fcd7aaac3fd5392ead09280ad01a33096b2a5681&f=plugins%2Fprogram_management%2Fdocs%2Fdecisions%2F0005-build-domain-objects.md

Change-Id: I07f9d50600bbed0a78b3ec477df9da6034cfd450