fix: Display errors and parent artifact in the modal
fixes request #41813 Errors and parent artifact are not shown in the
artifact modal

How to test:
- Go to Backlog, choose a User story in the backlog, click on its
shortname to open the dropdown, click on the menu to create a child item
(a Task for example). In the modal that opens, you should see at the top
an info alert that tells you the new artifact will be linked to the User
story you selected.
- Make network offline or forge the REST Resource class so that it
throws an error. When you submit the modal, the error message should be
in a danger alert at the top of the modal, with a translated message.

No functional change expected:
- when loading the comments (when editing an existing artifact)
- when loading the existing artifact links
- when there is an error while creating a new Artifact link from its
title
- with selectboxes that have field dependencies
- when showing the max File upload quota at the bottom left (need at
least one File field)

Probably introduced by f755f7e1d20aaf2ad392c900b26874cd383c4e36.

Why?
In Hybrids v8.x, you could describe the custom element's properties with
a getter and setter. We "abused" the setters by having them do
side-effects, such as triggering the loading of comments or registering
a callback to display errors.

With the upgrade to Hybrids v9.x, there is now only a single function
that does "getter and setter". This function will throw an error if it
detects that we are setting some unrelated properties, like we used to
do.

It is more correct to trigger side-effects in the "observe" function of
the properties. Since we only set the controller once and it is "cached"
later (even with AngularJS fuzziness), observe() will be called only
once per element, and everything will be alright.

There was a bug only for ModalFeedback due to "bad luck": since in the
other elements, the controller is called by another property, Hybrids
did trigger the "setter" function. It was only in ModalFeedback that the
controller is never called anywhere else in the element, so as Hybrids
is "lazy", it did not call it at all, hence no feedbacks.

Change-Id: I2c3280641af71afc6a3b73df91e0089b14f3021f