refactor: Composition API in ToBePlanned
part of request #41457 Migrate plugin-program-management-backlog to Vue
3

How to test:
No functional change.
When browsing a Program, you see the list of Features on the left. You
can drag/drop to reorder it, or move a Feature to a Program Increment.

Why?
We cannot got through `useState` directly to get the features.It wraps
the state in a `readonly()`, but Vue 2 does not support readonly for
arrays. The only solution is to access the state from `useStore`
directly, to avoid the `readonly()` array. This solution is not intended
to stay as it is, it should be changed after the switch to Vue 3.

Change-Id: I8a518d8a0f710133a7ae0649682704f4d1242283