refactor: Composition API in App
part of request #41457 Migrate plugin-program-management-backlog to Vue
3

How to test:
No functional change.
When browsing a Program, when the sidebar is collapsed, the breadcrumbs
include the project's name and icon and shields etc. You can drag/drop
to move Features in the Backlog and to Program Increments.
When the Program is not configured, the administrator has a link to the
admin page. Non-admin also sees an empty state, but with no link.

Why?
We cannot go through `useState` directly to get the features. It wraps
the state in a `readonly()`, but Vue 2 does not support readonly for
arrays. The only solution is to access the state from `useStore`
directly, to avoid the `readonly()` array. This solution is not intended
to stay as it is, it should be changed after the switch to Vue 3.

Change-Id: Ifd5e10bb7c1a7fc9aaf3662283485a97a2fe39c3