feat: Update level of a freetext section
Update of an artifact section is not done through artidoc backend yet,
so we cannot update its level for now.

No functional changes for regular users.

If you hack the db to set a level for a freetext section, and update it
through artidoc UI, it will be reset to level=1
If you use the REST APPI to update a freetext section, you can set its
level [1, 3]

Part of story #41805: Management of titles

Change-Id: I6cebd194436ab9590fd16f3467c1db877c519919